[#17] fix: 스키마에 맞지 않은 필드 타입 변경
Long형을 Integer형으로 변경

+++ src/main/java/dabang/star/cafe/domain/manager/Manager.java
@@ -11,7 +11,7 @@ public class Manager {
 
     private Long id;
 
-    private Long officeId;
+    private Integer officeId;
 
     private String name;
